refactor : modify class structure for intutive and time efficient
Application class only call play method Because I think so, main method has burden should be low

Refactor(commit_id) : #b2c2b54
